requirements.txt
paho-mqtt==1.6.1
pydantic==1.10.7
PyYAML==6.0
schedule==1.2.0
typing_extensions==4.5.0
